This commit addresses: http://bugs.kde.org/show_bug.cgi?id=115656 and quite possibly, but not verified: http://bugs.kde.org/show_bug.cgi?id=115486
In the words of the associated patch from wirr:

Problem cs2() gives pointer to short (2 Bytes)
wchar_t is the same as Py_UNICODE (4 Bytes)
Casting short to Py_UNICODE causes it to be
tmp[i] + tmp[i+1] which is not correct.
This may be platform and distributation dependant,
but should work on every platform, beacuase wchar_t is
always bigger or equal than short.


moved k->keyPressed(m_key, m_input); to keyPressEvent.
isEmpty is checking if it is a real char not a modifier.
Enter-Key is working




svn path=/branches/KDE/3.5/kdeutils/superkaramba/; revision=477748