Patch from Alex to fix unicode and input box bugs:
If Py_UNICODE is 4 Bytes do the same as in the last patch.
If it is 2 Bytes then it's the same as unsigned short.
Now why the cast to (wchar_t*) in PyUnicode_FromWideChar?
The code seems to get compiled for both Py_UNICODE 4 and 2 Bytes
so the cast is necessary. Why PyUnicode_FromWideChar for 2Bytes still
wants wchar_t is not clear to me.

Dropped m_key. Problem is that the keyPressed-method is called before
the text in the LineEdit is changed. calling the membere function last 
solved it.  Also solved that a few keys are not recognised (because of
fast typing)


svn path=/branches/KDE/3.5/kdeutils/superkaramba/; revision=479298